+++ noodles-cram/src/record/convert.rs
@@ -1,26 +1,20 @@
 use std::{io, str};
 
 use noodles_bam::record::ReferenceSequenceId;
-use noodles_fasta as fasta;
 use noodles_sam::{
     self as sam,
     record::{Data, QualityScores, Sequence},
 };
 
-use crate::data_container::CompressionHeader;
-
-use super::{
-    resolve::{resolve_bases, resolve_features},
-    Record, Tag,
-};
+use super::{resolve::resolve_features, Record, Tag};
 
 impl Record {
     /// Converts this CRAM record to a SAM record.
+    ///
+    /// This assumes this record is fully resolved.
     pub fn try_into_sam_record(
         &self,
-        reference_assembly: &[fasta::Record],
         reference_sequences: &sam::header::ReferenceSequences,
-        compression_header: &CompressionHeader,
     ) -> io::Result<sam::Record> {
         let mut builder = sam::Record::builder();
 
@@ -68,28 +62,7 @@ impl Record {
         builder = builder.set_template_length(self.template_size());
 
         if self.read_length() > 0 {
-            let sequence = if let Some(reference_sequence_id) = self.reference_sequence_id() {
-                let reference_sequence_record = usize::try_from(i32::from(reference_sequence_id))
-                    .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-                    .and_then(|id| {
-                        reference_assembly.get(id).ok_or_else(|| {
-                            io::Error::new(io::ErrorKind::InvalidData, "missing reference sequence")
-                        })
-                    })?;
-
-                let raw_bases = resolve_bases(
-                    reference_sequence_record,
-                    compression_header,
-                    self.features(),
-                    self.alignment_start().map(i32::from).unwrap_or_default(),
-                    self.read_length() as usize,
-                );
-
-                bytes_to_sequence(&raw_bases)?
-            } else {
-                bytes_to_sequence(self.bases())?
-            };
-
+            let sequence = bytes_to_sequence(self.bases())?;
             builder = builder.set_sequence(sequence);
         }